Add files via upload
Cryptography is the most effective and common technique used to achieve confidentiality, authentication and non-repudiation of data. In cryptography MACs are used between two parties that share a secret to authenticate information exchange between those parties. Galois/Counter Mode (GCM) is an algorithm for authenticated encryption proposed by National Institute of Standards and Technology (NIST). GCM is one of the secure and efficient authenticated technique. The objective of this project is to analyze the efficiency of GCM using MPI. Some recommendations based on experimental data for selection of efficient GCM for an end user.

It is a implementation of the paper:
https://ieeexplore.ieee.org/document/7058525